+++ src/main/java/groom/him/domain/search/repository/SearchRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SearchRepository extends JpaRepository<SearchEntity,Integer> {
+public interface SearchRepository extends JpaRepository<SearchEntity, Long> {
 }
